requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
discord==1.7.3
discord.py==1.7.3
idna==2.10
lxml==4.6.3
multidict==5.1.0
numpy==1.21.0
packaging==20.9
pandas==1.2.5
pip-review==1.1.0
praw==7.3.0
prawcore==2.2.0
pycodestyle==2.7.0
pyparsing==2.4.7
PyQt3D==5.15.4
PyQt5==5.15.4
PyQt5-sip==12.9.0
PyQtChart==5.15.4
PyQtDataVisualization==5.15.4
PyQtNetworkAuth==5.15.4
PyQtPurchasing==5.15.4
PyQtWebEngine==5.15.4
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
QScintilla==2.12.1
requests==2.25.1
schedule==1.1.0
scipy==1.7.0
six==1.16.0
soupsieve==2.2.1
toml==0.10.2
typing-extensions==3.10.0.0
update-checker==0.18.0
urllib3==1.26.6
websocket-client==1.1.0
yarl==1.6.3
